refactor: Define separate routers
