+++ .eslintrc.yaml
@@ -18,6 +18,6 @@ overrides:
   - files: ['test/setup.e2e.ts']
     rules:
       '@typescript-eslint/no-unused-vars': off
-  - files: ['src/app.ts', 'test/e2e/whitelist-routes.test.ts']
+  - files: ['test/e2e/whitelist-routes.test.ts', 'src/controllers/**/*.ts']
     rules:
       '@typescript-eslint/no-misused-promises': off
