+++ package.json
@@ -28,6 +28,7 @@
     "@types/express": "^4.17.15",
     "@types/jest": "^29.2.3",
     "@types/node": "^18.11.18",
+    "@types/serve-favicon": "^2.5.3",
     "@types/supertest": "^2.0.12",
     "@types/swagger-ui-express": "^4.1.3",
     "@types/yargs": "^17.0.19",
@@ -80,6 +81,7 @@
     "is-docker": "^3.0.0",
     "module-alias": "^2.2.2",
     "reflect-metadata": "^0.1.13",
+    "serve-favicon": "^2.5.0",
     "swagger-ui-express": "^4.6.0",
     "tslib": "^2.5.0",
     "typedi": "^0.10.0"
