+++ src/controllers/index.ts
@@ -2,5 +2,6 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from './api-docs'
 export * from './User/index'
 export * from './Whitelist/index'
